+++ R/h5p.R
@@ -16,6 +16,12 @@
 #' @param h5p.img The image to display in front of the toc entry
 #' @param h5p.link The link when the image is clicked (sends to an help page
 #' about learnr tutorials).
+#' @param icourse The course identifier, e.g., `"MATH101"`.
+#' @param institution The institution name, e.g., `"My University"`.
+#' @param acad_year The academic year, e.g., `"2023-2024"`.
+#' @param term The term, e.g., `"Q1"`.
+#' @param set The set identifier, e.g., `"21M"` where 21 is the year and M is a
+#' set identifier.
 #'
 #' @return HTML code that generates the iframe. It is most conveniently used
 #' inside and R inline expression in your R Markdown document on its own line
@@ -31,7 +37,8 @@
 h5p <- function(id, baseurl,
     idurl = "wp-admin/admin-ajax.php?action=h5p_embed&id=",
     width = 780, height = 500, toc = "", toc.def = "H5P exercise {id}",
-    h5p.img = "images/list-h5p.png", h5p.link = "h5p") {
+    h5p.img = "images/list-h5p.png", h5p.link = "h5p", icourse = "",
+    institution = "", acad_year = "", term = "", set = "") {
   if (!is.null(toc)) {
     # Add an entry in the ex_toc
     ex_toc <- getOption("learnitdown_ex_toc", "")
@@ -42,6 +49,34 @@ h5p <- function(id, baseurl,
     ex_toc <- paste0(ex_toc, "\n",
       glue::glue("- [![h5p]({h5p.img})]({h5p.link}) [{toc}](#h5p_{id})"))
     options(learnitdown_ex_toc = ex_toc)
+
+    # Also add an entry in the apps
+    apps <- getOption("learnitdown_apps", data.frame())
+    app <- data.frame(
+      app         = id,
+      type        = "h5p",
+      icourse     = icourse,
+      institution = institution,
+      course      = substring(id, 1, 1),
+      acad_year   = acad_year,
+      term        = term,
+      module      = substring(id, 1, 3),
+      set         = set,
+      assignment  = NA_character_,
+      template    = NA_character_,
+      url         = paste0(baseurl, "/", idurl, id),
+      alt_url     = paste0(baseurl, "/", idurl, id),
+      start       = NA,
+      end         = NA,
+      deadline    = NA,
+      part        = NA,
+      toc         = TRUE,
+      clone       = FALSE,
+      n           = 1,
+      level       = 1,
+      weight      = 1) # Always 1 for now
+    apps <- rbind(apps, app)
+    options(learnitdown_apps = apps)
   }
   # In H5P wordpress plugin 1.15.2, there is title="' . $title . '" inside the iframe attributes that is added
   glue::glue("\n[]{{#h5p_{id}}}[![h5p]({h5p.img})]({h5p.link})\n<iframe src=\"{baseurl}/{idurl}{id}\" width=\"{width}\" height=\"{height}\" frameborder=\"0\" allowfullscreen=\"allowfullscreen\" class=\"h5p\"></iframe><script src=\"{baseurl}/wp-content/plugins/h5p/h5p-php-library/js/h5p-resizer.js\" charset=\"UTF-8\"></script>\n")